+++ src/components/network/NetworkStatus.vue
